refactor: type season data normalizer
